+++ client/vscode/src/settings/accessTokenSetting.ts
@@ -9,25 +9,28 @@ export function accessTokenSetting(): string | undefined {
     return readConfiguration().get<string>('accessToken')
 }
 
+export async function removeAccessTokenSetting(): Promise<void> {
+    await readConfiguration().update('accessToken', 'REDACTED', vscode.ConfigurationTarget.Global)
+    return
+}
+
 // Ensure that only one access token error message is shown at a time.
 let showingAccessTokenErrorMessage = false
 
-export async function handleAccessTokenError(badToken?: string, endpointURL?: string): Promise<void> {
-    const currentValue = readConfiguration().get<string>('accessToken')
-
-    if (currentValue === badToken && !showingAccessTokenErrorMessage) {
+export async function handleAccessTokenError(badToken: string, endpointURL: string): Promise<void> {
+    if (badToken !== undefined && !showingAccessTokenErrorMessage) {
         showingAccessTokenErrorMessage = true
 
         const message = !badToken
             ? `A valid access token is required to connect to ${endpointURL}`
             : `Connection to ${endpointURL} failed because the token is invalid. Please reload VS Code if your Sourcegraph instance URL has changed.`
 
-        const version = await observeInstanceVersionNumber().toPromise()
+        const version = await observeInstanceVersionNumber(badToken, endpointURL).toPromise()
         const supportsTokenCallback = version && isOlderThan(version, { major: 3, minor: 41 })
-        const action = await vscode.window.showErrorMessage(message, 'Get Token', 'Open Settings')
+        const action = await vscode.window.showErrorMessage(message, 'Get Token', 'Update Instance URL in Setting')
 
         if (action === 'Open Settings') {
-            await vscode.commands.executeCommand('workbench.action.openSettings', 'sourcegraph.accessToken')
+            await vscode.commands.executeCommand('workbench.action.openSettings', 'sourcegraph.url')
         } else if (action === 'Get Token') {
             const path = supportsTokenCallback ? '/user/settings/tokens/new/callback' : '/user/settings/'
             const query = supportsTokenCallback ? 'requestFrom=VSCEAUTH' : ''
@@ -42,15 +45,5 @@ export async function handleAccessTokenError(badToken?: string, endpointURL?: st
             )
         }
         showingAccessTokenErrorMessage = false
-    }
-}
-
-export async function updateAccessTokenSetting(newToken: string): Promise<boolean> {
-    // TODO: STORE TOKEN IN KEYCHAIN AND REMOVE FROM USER CONFIG
-    try {
-        await readConfiguration().update('accessToken', newToken, vscode.ConfigurationTarget.Global)
-        return true
-    } catch {
-        return false
     }
 }